# contract/openapi.yaml
openapi: 3.0.3
info:
  title: BuryMe API
  version: 1.0.0
  description: |
    Backend contract for **BuryMe** — a peer-to-peer IOU / micro-lending tracker.
    Derived directly from PRD v2.7, §5 (Conceptual Model), §6–7 (Requirements &
    Scenarios), §8 (Entity Attribute Reference), §11 (Dependencies & Integrations),
    and §12 (Validation & Error Handling).

    ## Architecture
    - **Thin clients.** All business logic, persistence, and third-party integration
      live in the Express/Node backend (§11.0). React (web) and Flutter (mobile)
      consume this API only — neither has direct DB or third-party access.
    - **Auth is Firebase-backed** (§11.1.2). Phone/email OTP and credential
      management happen entirely client-side via the Firebase Auth SDK. The client
      sends the resulting Firebase ID token as `Authorization: Bearer <token>` on
      every request; Express verifies it via the Firebase Admin SDK. There is
      therefore no `/login`, `/send-otp`, or `/verify-otp` endpoint here — those
      never touch this API.
    - **First login provisions a BuryMe user.** A valid Firebase token with no
      matching BuryMe profile yields `404` on `GET /auth/me`; the client then calls
      `POST /auth/register`.

    ## Request vs. Obligation — the core modeling decision
    Per §5 and §8.5/§8.6, **Request** and **Obligation** are two distinct persisted
    resources, not two states of one resource:
...
      It never re-enters a negotiation state.

...
    pre-filled.
...
servers:
...
security:
  - firebaseIdToken: []
paths:
...
  /requests:
    get:
      tags: [Requests]
      summary: List requests involving the current user
      operationId: listRequests
      parameters:
        - name: role
          in: query
          description: Filter by the current user's role on the request.
          schema: { type: string, enum: [initiator, recipient] }
        - name: status
          in: query
          description: Comma-separated RequestStatus values.
          schema: { type: string }
        - name: request_type
          in: query
          schema: { $ref: "#/components/schemas/RequestType" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Page of requests.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RequestPage" }

    post:
      tags: [Requests]
      summary: Create a Borrow, Lend, or Repayment request
      description: >
        Body shape is discriminated by `request_type`. See
        `BorrowRequestInput` / `LendRequestInput` / `RepaymentRequestInput`.
        §12.2.2 validation applies per type. A user cannot target themselves, and
        Borrow/Lend cannot be created while an open request already exists between
        the same two users in the same direction (§7.1.3, §12.2.2).
      operationId: createRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/BorrowRequestInput"
                - $ref: "#/components/schemas/LendRequestInput"
                - $ref: "#/components/schemas/RepaymentRequestInput"
              discriminator:
                propertyName: request_type
                mapping:
                  Borrow: "#/components/schemas/BorrowRequestInput"
                  Lend: "#/components/schemas/LendRequestInput"
                  Repayment: "#/components/schemas/RepaymentRequestInput"
      responses:
        "201":
          description: Request created with status `Pending`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409":
          description: An open request already exists, or the Repayment target isn't actually overdue.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /requests/{requestId}:
    get:
      tags: [Requests]
      summary: Get a request
      operationId: getRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "404": { $ref: "#/components/responses/NotFound" }
  /requests/{requestId}/counter:
    post:
      tags: [Requests]
      summary: Submit the single allowed counter-proposal
      description: >
        Negotiation is capped at two proposals total (§6.2.1): the original, then
        one counter. Allowed only while `status = Pending`, and only once — a
        second counter attempt returns `409 STATUS_CONFLICT`. After a counter, only
        the original initiator may accept or decline; the counter-submitter cannot
        counter again.
      operationId: counterRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CounterProposalInput" }
      responses:
        "200":
          description: Request updated with the counter-proposal, status `Countered`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "409": { $ref: "#/components/responses/StatusConflict" }
  /requests/{requestId}/accept:
    post:
      tags: [Requests]
      summary: Accept the current terms
      description: >
        For Borrow/Lend: creates the Obligation and sets `request.obligation_id`.
        The obligation enters `Active` immediately if `disbursement_method =
        Already Given`, or `Pending Disbursement` if `Through App` (§6.3.1). For
        Borrow specifically, `disbursement_method` is supplied by the lender at
        acceptance (§8.5) — see `AcceptBorrowRequestInput`. For Repayment: marks
        the request `Accepted`; the actual balance update happens when the
        resulting payment is confirmed, not here.
      operationId: acceptRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptRequestInput" }
      responses:
        "200":
          description: Updated request (now `Accepted`), with `obligation_id` populated for Borrow/Lend.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "409": { $ref: "#/components/responses/StatusConflict" }
  /requests/{requestId}/decline:
    post:
      tags: [Requests]
      summary: Decline the request
      operationId: declineRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request moved to `Declined`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "409": { $ref: "#/components/responses/StatusConflict" }
  /requests/{requestId}/cancel:
    post:
      tags: [Requests]
      summary: Withdraw a request (initiator only, while Pending)
      operationId: cancelRequest
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request moved to `Cancelled`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Request" }
        "403": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/StatusConflict" }

  # ══════════════════════════════════════════════════════════════
  # OBLIGATIONS
  # ══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Firebase ID token (1-hour expiry, §11.1.2). Verified on the backend via the
        Firebase Admin SDK on every protected route.
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    ValidationError:
      description: One or more input fields failed validation (§12.3.2 `VALIDATION_ERROR`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_ERROR
              message: The due date must be a future date.
              field: proposed_due_date
...
    Unauthorized:
      description: Authenticated but not a party to this record (§12.3.2 `UNAUTHORIZED`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Record does not exist or is not accessible to this user (§12.3.2 `NOT_FOUND`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    StatusConflict:
      description: The record's current status does not allow this operation (§12.3.2 `STATUS_CONFLICT`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    ErrorCode:
      type: string
      description: Standard error codes per §12.3.2.
      enum:
        - VALIDATION_ERROR
        - DUPLICATE_SUBMISSION
        - STATUS_CONFLICT
        - BALANCE_EXCEEDED
        - CONCURRENT_MODIFICATION
        - PAYMENT_GATEWAY_UNAVAILABLE
        - PAYMENT_GATEWAY_ERROR
        - RECIPIENT_UNVERIFIED
        - PARTIAL_WRITE_ROLLED_BACK
        - UNAUTHENTICATED
        - UNAUTHORIZED
        - NOT_FOUND
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Plain-language, user-safe. Never exposes stack traces, DB ids, or raw third-party payloads (§12.1).
            field:
              type: string
              nullable: true
              description: Present only for field-level validation errors; omitted for system errors (§12.3.1).

    # ── Enums (verbatim from §8) ─────────────────────────────────
    RequestType:
      type: string
      enum: [Borrow, Lend, Repayment]
    RequestStatus:
      type: string
      enum: [Pending, Countered, Accepted, Declined, Cancelled]
    RepaymentType:
      type: string
      enum: [Lump Sum, Installments]
    DisbursementMethod:
      type: string
      enum: [Already Given, Through App]
...
    Money:
      type: object
      description: All amounts in ETB, ≤2 decimal places (§12.1).
      required: [amount, currency]
      properties:
        amount: { type: number, format: double, minimum: 0.01, maximum: 1000000, multipleOf: 0.01 }
        currency: { type: string, default: ETB, enum: [ETB] }

    # ── Auth / Users ──────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [user_id, display_name, profile_photo_url]
      properties:
        user_id: { type: string }
        display_name: { type: string }
        profile_photo_url: { type: string, nullable: true }
...
    ProposedScheduleInput:
      type: object
      description: §12.2.3 — required iff proposed_repayment_type is Installments; 2–60 entries; sum must equal the principal (0.01 ETB tolerance); dates ascending, future, ≤ overall due date.
      required: [installments]
      properties:
        installments:
          type: array
          minItems: 2
          maxItems: 60
          items:
            type: object
            required: [amount, due_date]
            properties:
              amount: { $ref: "#/components/schemas/Money" }
              due_date: { type: string, format: date }
    BorrowRequestInput:
      type: object
      description: §6.2.1, §12.2.2. Lender sets disbursement_method later, at acceptance.
      required:
        [
          request_type,
          recipient_user_id,
          amount,
          purpose,
          proposed_repayment_type,
          proposed_due_date,
        ]
      properties:
        request_type: { type: string, enum: [Borrow] }
        recipient_user_id: { type: string, description: The prospective lender. }
        amount: { $ref: "#/components/schemas/Money" }
        purpose: { type: string, minLength: 3, maxLength: 200 }
        proposed_repayment_type: { $ref: "#/components/schemas/RepaymentType" }
        proposed_schedule: { $ref: "#/components/schemas/ProposedScheduleInput" }
        proposed_due_date:
          type: string
          format: date
          description: ≥1 day in the future, ≤3 years from today (§12.2.2).
      additionalProperties: false
    LendRequestInput:
      type: object
      description: §6.2.4, §12.2.2. Lender sets disbursement_method at creation.
      required:
        [
          request_type,
          recipient_user_id,
          amount,
          purpose,
          proposed_repayment_type,
          proposed_due_date,
          disbursement_method,
        ]
      properties:
        request_type: { type: string, enum: [Lend] }
        recipient_user_id: { type: string, description: The borrower. }
        amount: { $ref: "#/components/schemas/Money" }
        purpose: { type: string, minLength: 3, maxLength: 200 }
        proposed_repayment_type: { $ref: "#/components/schemas/RepaymentType" }
        proposed_schedule: { $ref: "#/components/schemas/ProposedScheduleInput" }
        proposed_due_date: { type: string, format: date }
        disbursement_method: { $ref: "#/components/schemas/DisbursementMethod" }
      additionalProperties: false
    RepaymentRequestBody:
      type: object
      description: Shared body for repayment-request creation, used directly by the convenience endpoint and nested in RepaymentRequestInput.
      required: [target]
      properties:
        installment_id:
          type: string
          nullable: true
          description: Required (must reference a currently-Overdue installment) for Installments obligations; omit for Lump Sum.
        note: { type: string, maxLength: 300, nullable: true }
      additionalProperties: false
    RepaymentRequestInput:
      allOf:
        - type: object
          required: [request_type, obligation_id]
          properties:
            request_type: { type: string, enum: [Repayment] }
            obligation_id:
              type: string
              description: Set immediately at creation (unlike Borrow/Lend, where obligation_id is set only on acceptance).
        - $ref: "#/components/schemas/RepaymentRequestBody"
    CounterProposalInput:
      type: object
      description: Same shape/validation as the original proposal for this request's type.
      properties:
        amount: { $ref: "#/components/schemas/Money" }
        proposed_repayment_type: { $ref: "#/components/schemas/RepaymentType" }
        proposed_schedule: { $ref: "#/components/schemas/ProposedScheduleInput" }
        proposed_due_date: { type: string, format: date }
    AcceptRequestInput:
      type: object
      properties:
        disbursement_method:
          allOf: [{ $ref: "#/components/schemas/DisbursementMethod" }]
          description: Required only when accepting a Borrow request (the lender chooses it here); ignored otherwise.
    Request:
      type: object
      required:
        [
          request_id,
          request_type,
          initiating_user,
          receiving_user,
          obligation_id,
          amount,
          purpose,
          proposed_repayment_type,
          proposed_schedule,
          proposed_due_date,
          disbursement_method,
          counter_proposal,
          status,
          created_at,
          responded_at,
        ]
      properties:
        request_id: { type: string }
        request_type: { $ref: "#/components/schemas/RequestType" }
        initiating_user: { $ref: "#/components/schemas/UserSummary" }
        receiving_user: { $ref: "#/components/schemas/UserSummary" }
        obligation_id:
          type: string
          nullable: true
          description: Null until acceptance for Borrow/Lend; set immediately for Repayment.
        amount: { $ref: "#/components/schemas/Money" }
        purpose: { type: string, nullable: true }
        proposed_repayment_type:
          allOf: [{ $ref: "#/components/schemas/RepaymentType" }]
          nullable: true
        proposed_schedule:
          allOf: [{ $ref: "#/components/schemas/ProposedScheduleInput" }]
          nullable: true
        proposed_due_date: { type: string, format: date, nullable: true }
        disbursement_method:
          allOf: [{ $ref: "#/components/schemas/DisbursementMethod" }]
          nullable: true
        counter_proposal:
          allOf: [{ $ref: "#/components/schemas/CounterProposalInput" }]
          nullable: true
          description: Populated once status is Countered.
        status: { $ref: "#/components/schemas/RequestStatus" }
        created_at: { type: string, format: date-time }
        responded_at: { type: string, format: date-time, nullable: true }
    RequestPage:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Request" }
        next_cursor: { type: string, nullable: true }

    # ── Obligations (§8.6) ────────────────────────────────────────
...